+++ .dockerignore
@@ -1,0 +1,5 @@
+./bin
+./log/*
+./.git
+./README.md
+docker/
