+++ README.md
@@ -1,5 +1,4 @@
 TODOS:
 
 - swagger spec
-- dockerizing
 - tests
