+++ docker/docker-compose.mysql.yml
@@ -1,0 +1,11 @@
+version: "3.4"
+
+services:
+  db:
+    image: mysql:${MYSQL_IMAGE_TAG}
+    container_name: yorpoll_mysql
+    environment:
+      - MYSQL_ROOT_PASSWORD=${DATABASE_ROOT_PASSWORD}
+      - MYSQL_DATABASE=${DATABASE_NAME} 
+      - MYSQL_USER=${DATABASE_USER}
+      - MYSQL_PASSWORD=${DATABASE_PASSWORD}
