+++ docker/docker-compose.server.yml
@@ -1,0 +1,30 @@
+version: "3.4"
+
+services:
+  app:
+    build:
+      dockerfile: ./Dockerfile.server
+      context: ../
+      target: app
+    container_name: yorpoll_app
+    volumes:
+      - type: bind
+        source: ../logs
+        target: /app/logs
+    environment:
+      - SERVER_PORT=${SERVER_PORT}
+      - SERVER_HOST=0.0.0.0
+      - DATABASE_TYPE=${DATABASE_TYPE}
+      - DATABASE_NAME=${DATABASE_NAME}
+      - DATABASE_PORT=${DATABASE_PORT}
+      - DATABASE_HOST=db
+      - DATABASE_USER=${DATABASE_USER}
+      - DATABASE_PASSWORD=${DATABASE_PASSWORD}
+      - ENV=${ENV}
+      - LOG_FILE=${LOG_FILE}
+    ports:
+      - "${SERVER_PORT}:${SERVER_PORT}"
+    depends_on:
+      - db
+    restart: always
+
