+++ internal/db/mongodb/mongodb.go
@@ -55,6 +55,12 @@ func (c *Connection) Connect(config *config.Config) error {
 		return err
 	}
 
+	err = client.Ping(ctx, readpref.Primary())
+	if err != nil {
+		log.Error(err)
+		return err
+	}
+
 	c.client = client
 	return nil
 }
